+++ odoo/local-src/stock_valuation/__manifest__.py
@@ -11,6 +11,7 @@
         'stock_account',
         'specific_purchase',  # for product supplier
         'specific_base',
+        'product_last_transaction',
     ],
     'data': [
         'wizards/stock_valuation_history_view.xml',